Add files via upload
This is the first upload involving the MVC pattern.
I got the physics working with the MVC pattern. Left and right to move and space to shoot are there with the ball bouncing around the map. I have to program the bricks and map loading tomorrow, because those involve statements of their own. 

The main method is found in controller. DO NOT use the main method in Window.